Start working on 0.8.0 (ring.logger revamp)
First step towards ring.logger 0.8.0, a major revamp of the library.
The goal is to simplify the library to make it more easy to use and
adapt.

Uses basic clojure data structures instead of strings as
messages. Replaces the Logger protocol with a log-fn that defaults to
`clojure.tools.logging/log`. Gets rid of the multimethods in
ring.logger/messages in favor of leaving it to the user to format
messages in the log-fn (and maybe in an output-fn in the future).